# docs/api/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/projects":{"get":{"operationId":"ProjectsController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"example":1,"type":"string"}},{"name":"per_page","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"string"}},{"name":"search","required":false,"in":"query","description":"Search term","schema":{"type":"string"}},{"name":"is_featured","required":false,"in":"query","description":"Filter by featured projects","schema":{"type":"boolean"}}],"responses":{"200":{"description":"List of projects","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectsListResponseDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get all projects","tags":["Projects"]},"post":{"operationId":"ProjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":"The project has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleProjectResponseDto"}}}},"422":{"description":"Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new project","tags":["Projects"]}},"/projects/{id}":{"get":{"operationId":"ProjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Project ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The found project","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleProjectResponseDto"}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get a project by ID","tags":["Projects"]},"patch":{"operationId":"ProjectsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Project ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"The project has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleProjectResponseDto"}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"422":{"description":"Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update a project by ID","tags":["Projects"]},"delete":{"operationId":"ProjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Project ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The project has been successfully deleted.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResponseDto"}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Delete a project by ID","tags":["Projects"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"The record found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"},"example":{"status":"error","error":{"code":"AUTHENTICATION_ERROR","message":"Password does not match"}}}}},"422":{"description":"Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"User Login","tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"The record found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"422":{"description":"Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"User Registration","tags":["auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_profile","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get User Profile","tags":["auth"]}}},"info":{"title":"Portfolio API","description":"API for my personal portfolio","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ProjectResponseDto":{"type":"object","properties":{"id":{"type":"number","description":"Project ID"},"title":{"type":"string","description":"Project title"},"description":{"type":"string","description":"Project description"},"shortDescription":{"type":"string","description":"Short description"},"liveUrl":{"type":"string","description":"Live URL"},"repoUrl":{"type":"string","description":"Repository URL"},"isFeatured":{"type":"boolean","description":"Is the project featured?"},"createdAt":{"format":"date-time","type":"string","description":"Creation date"},"updatedAt":{"format":"date-time","type":"string","description":"Last update date"}},"required":["id","title","description","shortDescription","repoUrl","isFeatured","createdAt","updatedAt"]},"PaginationMetaDto":{"type":"object","properties":{"page":{"type":"number","description":"Current page number","example":1},"limit":{"type":"number","description":"Number of items per page","example":10},"totalItems":{"type":"number","description":"Total number of items","example":100},"totalPages":{"type":"number","description":"Total number of pages","example":10}},"required":["page","limit","totalItems","totalPages"]},"ResponseMetaDto":{"type":"object","properties":{"pagination":{"description":"Pagination metadata","allOf":[{"$ref":"#/components/schemas/PaginationMetaDto"}]}}},"ProjectsListResponseDto":{"type":"object","properties":{"status":{"type":"string","description":"Status of the response","enum":["success","error"],"example":"success"},"data":{"description":"List of projects","type":"array","items":{"$ref":"#/components/schemas/ProjectResponseDto"}},"meta":{"description":"Response metadata","allOf":[{"$ref":"#/components/schemas/ResponseMetaDto"}]}},"required":["status","data","meta"]},"ErrorDetailsDto":{"type":"object","properties":{}},"ErrorInfoDto":{"type":"object","properties":{"code":{"type":"string","description":"Error code","enum":["VALIDATION_ERROR","AUTHENTICATION_ERROR","AUTHORIZATION_ERROR","RESOURCE_NOT_FOUND","INTERNAL_SERVER_ERROR"],"example":"VALIDATION_ERROR"},"message":{"type":"string","description":"Human-readable error message","example":"Validation failed"},"details":{"description":"Additional error details","allOf":[{"$ref":"#/components/schemas/ErrorDetailsDto"}]}},"required":["code","message"]},"ErrorResponseDto":{"type":"object","properties":{"status":{"type":"string","description":"Status of the response","enum":["success","error"],"example":"success"},"error":{"description":"Error information","allOf":[{"$ref":"#/components/schemas/ErrorInfoDto"}]}},"required":["status","error"]},"SingleProjectResponseDto":{"type":"object","properties":{"status":{"type":"string","description":"Status of the response","enum":["success","error"],"example":"success"},"data":{"description":"Project data","allOf":[{"$ref":"#/components/schemas/ProjectResponseDto"}]},"meta":{"description":"Response metadata","allOf":[{"$ref":"#/components/schemas/ResponseMetaDto"}]}},"required":["status","data","meta"]},"CreateProjectDto":{"type":"object","properties":{"title":{"type":"string","description":"Project title"},"description":{"type":"string","description":"Project description"},"shortDescription":{"type":"string","description":"Short description"},"liveUrl":{"type":"string","description":"Live URL"},"repoUrl":{"type":"string","description":"Repository URL"},"isFeatured":{"type":"boolean","description":"Is the project featured?"}},"required":["title","description","shortDescription","repoUrl"]},"UpdateProjectDto":{"type":"object","properties":{"title":{"type":"string","description":"Project title"},"description":{"type":"string","description":"Project description"},"shortDescription":{"type":"string","description":"Short description"},"liveUrl":{"type":"string","description":"Live URL"},"repoUrl":{"type":"string","description":"Repository URL"},"isFeatured":{"type":"boolean","description":"Is the project featured?"}}},"DeleteResponseDto":{"type":"object","properties":{"status":{"type":"string","description":"Status of the response","enum":["success","error"],"example":"success"},"data":{"type":"object","description":"Response message","example":{"message":"Project successfully deleted"}},"meta":{"description":"Response metadata","allOf":[{"$ref":"#/components/schemas/ResponseMetaDto"}]}},"required":["status","data","meta"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User Email"},"password":{"type":"string","description":"User Password"}},"required":["email","password"]},"SuccessResponseDto":{"type":"object","properties":{"status":{"type":"string","description":"Status of the response","enum":["success","error"],"example":"success"},"data":{"type":"object","description":"Response data"},"meta":{"description":"Response metadata","allOf":[{"$ref":"#/components/schemas/ResponseMetaDto"}]}},"required":["status","data","meta"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"Username"},"email":{"type":"string","description":"User Email"},"password":{"type":"string"}},"required":["username","email","password"]}}}}